+++ Backend-AuraNeuro.API/Patient/Domain/Command/DeactivatePatientCommand.cs
@@ -3,4 +3,4 @@ namespace Backend_AuraNeuro.API.Patient.Domain.Command;
 /// <summary>
 /// Command to deactivate (soft delete) a patient.
 /// </summary>
-public record DeactivatePatientCommand(int PatientId);
+public record DeactivatePatientCommand(long PatientId);
